backend/requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.15,<4.0
djangorestframework-simplejwt>=5.3,<6.0
Pillow>=10.0,<11.0
django-cors-headers>=4.4,<5.0
